Restored SFBasic.java and moved quests there
